test: relocate the compaction cases to agent/context-compaction.test.ts
Second commit of PR 3. Relocation, not a rewrite: the three cases keep every
assertion, still construct a real AgentService through the harness, and still
read the JSONL protocol log rather than the controller.

That is the whole point for these three. The first asserts the wire order
["turn/start", "thread/compact/start", "turn/start"] and that exactly two
turn-started events reach the renderer, which is what proves the compaction turn
is swallowed rather than shown. The second pins the 80% threshold through the
79000/82000 pair against the fake's context window. The third drives
OPENBOT_FAKE_COMPACTION_ERROR and asserts queued work still runs. As unit tests
over ContextCompaction they would assert the mock, not the behaviour.

Tests: agent/context-compaction.test.ts 3 passed,
agent-service.queue.test.ts 23 passed; the same 26 the shard ran before the
split. tsc on tsconfig.node.json and biome on both files clean.
