+++ src/backend/agent-service.queue.test.ts
@@ -883,64 +883,6 @@ describe.sequential("AgentService: queue", () => {
     expect(["failed", "interrupted"]).toContain(service.listQueue("chief").deliveries[0]?.status);
   });
 
-  it("compacts a pressured agent context before draining its next queued message", async () => {
-    process.env.OPENBOT_FAKE_AUTO_COMPLETE = "DONE";
-    process.env.OPENBOT_FAKE_CONTEXT_USAGE = "82000";
-    const { store, mailbox } = stores(root);
-    service = new AgentService(store, mailbox, fakeBrowser());
-    const events: AgentEvent[] = [];
-    service.on("event", (event) => events.push(event));
-    await service.initialize();
-
-    await service.sendMessage({ botId: "chief", text: "First large task" });
-    await service.sendMessage({ botId: "chief", text: "Run after compaction" });
-    await waitFor(async () => {
-      const messages = await protocolMessages(logPath);
-      return messages.filter((message) => message.method === "turn/start").length === 2;
-    });
-
-    const lifecycle = (await protocolMessages(logPath))
-      .filter((message) => ["turn/start", "thread/compact/start"].includes(String(message.method)))
-      .map((message) => message.method);
-    expect(lifecycle.slice(0, 3)).toEqual(["turn/start", "thread/compact/start", "turn/start"]);
-    expect(lifecycle.filter((method) => method === "thread/compact/start")).toHaveLength(1);
-    expect(events.filter((event) => event.type === "turn-started")).toHaveLength(2);
-  });
-
-  it("does not compact context below the safety threshold", async () => {
-    process.env.OPENBOT_FAKE_AUTO_COMPLETE = "DONE";
-    process.env.OPENBOT_FAKE_CONTEXT_USAGE = "79000";
-    const { store, mailbox } = stores(root);
-    service = new AgentService(store, mailbox, fakeBrowser());
-    await service.initialize();
-    await service.sendMessage({ botId: "chief", text: "Normal task" });
-    await waitFor(() => service?.listQueue("chief").deliveries[0]?.status === "completed");
-
-    expect((await protocolMessages(logPath)).some((message) => message.method === "thread/compact/start")).toBe(false);
-  });
-
-  it("continues queued work when context compaction is unavailable", async () => {
-    process.env.OPENBOT_FAKE_AUTO_COMPLETE = "DONE";
-    process.env.OPENBOT_FAKE_CONTEXT_USAGE = "82000";
-    process.env.OPENBOT_FAKE_COMPACTION_ERROR = "1";
-    const { store, mailbox } = stores(root);
-    service = new AgentService(store, mailbox, fakeBrowser());
-    const events: AgentEvent[] = [];
-    service.on("event", (event) => events.push(event));
-    await service.initialize();
-
-    await service.sendMessage({ botId: "chief", text: "First task" });
-    await service.sendMessage({ botId: "chief", text: "Must still run" });
-    await waitFor(async () => {
-      const messages = await protocolMessages(logPath);
-      return messages.filter((message) => message.method === "turn/start").length === 2;
-    });
-
-    expect(events).toEqual(
-      expect.arrayContaining([expect.objectContaining({ type: "error", code: "context_compaction_failed" })]),
-    );
-  });
-
   it("fans out an idempotent agent tool message with referenced files", async () => {
     process.env.OPENBOT_FAKE_AGENT_TOOL = "1";
     const notePath = join(root, "generated-note.txt");
