+++ src/backend/agent/context-compaction.test.ts
@@ -1,0 +1,85 @@
+// @vitest-environment node
+import type { AgentEvent } from "@openbot/contracts/ipc";
+import { afterEach, beforeEach, describe, expect, it } from "vitest";
+import { AgentService } from "../agent-service";
+import {
+  fakeBrowser,
+  protocolMessages,
+  startAgentTestFixture,
+  stopAgentTestFixture,
+  stores,
+  waitFor,
+} from "../agent-service-test-harness";
+
+let root: string;
+let logPath: string;
+let service: AgentService | null = null;
+
+beforeEach(async () => {
+  ({ root, logPath } = await startAgentTestFixture());
+});
+
+afterEach(async () => {
+  await stopAgentTestFixture(root, service);
+  service = null;
+});
+
+describe.sequential("ContextCompaction: pressure, threshold and failure", () => {
+  it("compacts a pressured agent context before draining its next queued message", async () => {
+    process.env.OPENBOT_FAKE_AUTO_COMPLETE = "DONE";
+    process.env.OPENBOT_FAKE_CONTEXT_USAGE = "82000";
+    const { store, mailbox } = stores(root);
+    service = new AgentService(store, mailbox, fakeBrowser());
+    const events: AgentEvent[] = [];
+    service.on("event", (event) => events.push(event));
+    await service.initialize();
+
+    await service.sendMessage({ botId: "chief", text: "First large task" });
+    await service.sendMessage({ botId: "chief", text: "Run after compaction" });
+    await waitFor(async () => {
+      const messages = await protocolMessages(logPath);
+      return messages.filter((message) => message.method === "turn/start").length === 2;
+    });
+
+    const lifecycle = (await protocolMessages(logPath))
+      .filter((message) => ["turn/start", "thread/compact/start"].includes(String(message.method)))
+      .map((message) => message.method);
+    expect(lifecycle.slice(0, 3)).toEqual(["turn/start", "thread/compact/start", "turn/start"]);
+    expect(lifecycle.filter((method) => method === "thread/compact/start")).toHaveLength(1);
+    expect(events.filter((event) => event.type === "turn-started")).toHaveLength(2);
+  });
+
+  it("does not compact context below the safety threshold", async () => {
+    process.env.OPENBOT_FAKE_AUTO_COMPLETE = "DONE";
+    process.env.OPENBOT_FAKE_CONTEXT_USAGE = "79000";
+    const { store, mailbox } = stores(root);
+    service = new AgentService(store, mailbox, fakeBrowser());
+    await service.initialize();
+    await service.sendMessage({ botId: "chief", text: "Normal task" });
+    await waitFor(() => service?.listQueue("chief").deliveries[0]?.status === "completed");
+
+    expect((await protocolMessages(logPath)).some((message) => message.method === "thread/compact/start")).toBe(false);
+  });
+
+  it("continues queued work when context compaction is unavailable", async () => {
+    process.env.OPENBOT_FAKE_AUTO_COMPLETE = "DONE";
+    process.env.OPENBOT_FAKE_CONTEXT_USAGE = "82000";
+    process.env.OPENBOT_FAKE_COMPACTION_ERROR = "1";
+    const { store, mailbox } = stores(root);
+    service = new AgentService(store, mailbox, fakeBrowser());
+    const events: AgentEvent[] = [];
+    service.on("event", (event) => events.push(event));
+    await service.initialize();
+
+    await service.sendMessage({ botId: "chief", text: "First task" });
+    await service.sendMessage({ botId: "chief", text: "Must still run" });
+    await waitFor(async () => {
+      const messages = await protocolMessages(logPath);
+      return messages.filter((message) => message.method === "turn/start").length === 2;
+    });
+
+    expect(events).toEqual(
+      expect.arrayContaining([expect.objectContaining({ type: "error", code: "context_compaction_failed" })]),
+    );
+  });
+});
